requirements.txt
Flask==1.0.2
flask_restful==0.3.7
python-dotenv==0.10.1

mysql-connector-python==8.0.15
PyMySQL==0.9.3

requests==2.21.0

SQLAlchemy==1.3.3

Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
